requirements.txt
aiofiles==24.1.0
cssselect==1.1.0
parsel==1.6.0
Pillow==10.4.0
httpx==0.28.1
tenacity==8.2.2
pydantic==2.5.2
redis==4.6.0
openpyxl
pandas
cryptography
aiohttp
pyhumps==3.8.0
typer==0.16.0
loguru==0.7.2
PyExecJS==1.5.1
